Updated addFiles method to address JSON response
When there is a PHP error, response.response is an HTML string, not a JSON string. I set up a regex to test before parsing JSON. If it is an HTML string, I throw the $.ee_notice then return so it will log in the CP, not a JS error.